Improvements to skipping behavior.
Skipping now pauses on the first sound (bike bell ring) to indicate when series restart.
Volume is now reduced in advance of the first skip (first skip time /2) as that's usually the intent.

Also a master volume divisor was added to try to create line level output (as opposed to 0-VDC) but too much quality is lost.  It's disappointing that the DAC doesn't have built-in analog scaling.  Relatedly the suggested optional line level conversion resistors on the schematic are updated to 1K/1K, should give outputs 0-1.8V but is untested.